FIX var type int to int64

+++ CoinInfo.h
@@ -2,6 +2,7 @@
 #define COIN_INFO_H
 
 #include <ctime>
+#include <memory>
 
 namespace Gazua {
     class CoinInfo;
@@ -11,20 +12,20 @@ class Gazua::CoinInfo {
 public:
     //ticker
     std::time_t timestamp; //최종 체결 시각
-    unsigned int last,     //최종 체결 가격
-                open,      //최근 24시간 시작 가격
-                bid,       //매수호가
-                ask,       //매도호가
-                low,       //최근 24시간 체결 가격 중 최저가
-                high;      //최근 24시간 체결 가격 중 최고가
+    uint64_t last,         //최종 체결 가격
+            open,          //최근 24시간 시작 가격
+            bid,           //매수호가
+            ask,           //매도호가
+            low,           //최근 24시간 체결 가격 중 최저가
+            high;          //최근 24시간 체결 가격 중 최고가
     double volume;         //거래량
-    int change;            //시작 가격 대비 현재가 차이
+    int64_t change;        //시작 가격 대비 현재가 차이
     double changePercent;  //시작 가격 대비 현재가 차이 변화 비율
 
-   //constraint
-    unsigned int tick_size, //호가 단위
-                min_price,  //최소 주문가
-                max_price;  //최대 주문가
+    //constraint
+    uint64_t tick_size,     //호가 단위
+            min_price,      //최소 주문가
+            max_price;      //최대 주문가
     double order_min_size,  //매수/매도 수량 최소 입력값
         order_min_size;     //매수/매도 수량 최대 입력값            
 };
